gpt-app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
python-dotenv==1.0.0
google-auth==2.23.4
requests==2.31.0
python-multipart==0.0.6 